+++ src/globus_cli/commands/endpoint/_common.py
@@ -4,27 +4,10 @@ import typing as t
 
 import click
 
-from globus_cli import utils
 from globus_cli.constants import EXPLICIT_NULL
 from globus_cli.endpointish import EntityType
-from globus_cli.parsing import LocationType, MutexInfo, mutex_option_group
 
 C = t.TypeVar("C", bound=t.Union[t.Callable, click.Command])
-
-_GCSONLY = "(Globus Connect Server only)"
-_MANAGEDONLY = "(Managed endpoints only)"
-
-
-def endpoint_setattr_params(f: C) -> C:
-    return utils.fold_decorators(
-        f,
-        (
-            [_location_option, _public_option]
-            + _endpoint_network_use_params
-            + _endpoint_activation_params
-            + _endpoint_subscription_params
-        ),
-    )
 
 
 def validate_endpoint_create_and_update_params(
@@ -120,106 +103,3 @@ def validate_endpoint_create_and_update_params(
     if params.get("no_default_directory"):
         params["default_directory"] = EXPLICIT_NULL
         params.pop("no_default_directory")
-
-
-_endpoint_activation_params = [
-    click.option("--myproxy-dn", help=f"Set the MyProxy Server DN {_GCSONLY}"),
-    click.option("--myproxy-server", help=f"Set the MyProxy Server URI {_GCSONLY}"),
-    click.option("--oauth-server", help=f"Set the OAuth Server URI {_GCSONLY}"),
-]
-
-_endpoint_subscription_params = [
-    click.option(
-        "--managed",
-        "managed",
-        is_flag=True,
-        flag_value=True,
-        default=None,
-        help=(
-            "Set the endpoint as a managed endpoint. Requires the "
-            "user to be a subscription manager. If the user has "
-            "multiple subscription IDs, --subscription-id must be used "
-            "instead"
-        ),
-    ),
-    click.option(
-        "--no-managed",
-        "managed",
-        is_flag=True,
-        flag_value=False,
-        default=None,
-        help=(
-            "Unset the endpoint as a managed endpoint. "
-            "Does not require the user to be a subscription manager. "
-            "Mutually exclusive with --subscription-id"
-        ),
-    ),
-    click.option(
-        "--subscription-id",
-        type=click.UUID,
-        default=None,
-        help="Set the endpoint as a managed endpoint with the given "
-        "subscription ID. Mutually exclusive with --no-managed",
-    ),
-    mutex_option_group(
-        "--subscription-id",
-        MutexInfo(
-            "--no-managed", param="managed", present=lambda d: d.get("managed") is False
-        ),
-    ),
-]
-
-_endpoint_network_use_params = [
-    click.option(
-        "--network-use",
-        default=None,
-        type=click.Choice(["normal", "minimal", "aggressive", "custom"]),
-        help=(
-            "Set the endpoint's network use level. If using custom, "
-            "the endpoint's max and preferred concurrency and "
-            f"parallelism must be set {_MANAGEDONLY} {_GCSONLY}"
-        ),
-    ),
-    click.option(
-        "--max-concurrency",
-        type=int,
-        default=None,
-        help="Set the endpoint's max concurrency; requires --network-use=custom "
-        f"{_MANAGEDONLY} {_GCSONLY}",
-    ),
-    click.option(
-        "--preferred-concurrency",
-        type=int,
-        default=None,
-        help="Set the endpoint's preferred concurrency; requires --network-use=custom "
-        f"{_MANAGEDONLY} {_GCSONLY}",
-    ),
-    click.option(
-        "--max-parallelism",
-        type=int,
-        default=None,
-        help="Set the endpoint's max parallelism; requires --network-use=custom "
-        f"{_MANAGEDONLY} {_GCSONLY}",
-    ),
-    click.option(
-        "--preferred-parallelism",
-        type=int,
-        default=None,
-        help="Set the endpoint's preferred parallelism; requires --network-use=custom "
-        f"{_MANAGEDONLY} {_GCSONLY}",
-    ),
-]
-
-_public_option = click.option(
-    "--public/--private",
-    "public",
-    default=None,
-    help=f"Set the endpoint to be public or private {_GCSONLY}",
-)
-
-_location_option = click.option(
-    "--location",
-    type=LocationType(),
-    default=None,
-    help=f"Manually set the endpoint's latitude and longitude {_GCSONLY}",
-)
