+++ src/globus_cli/commands/gcp/create/mapped.py
@@ -4,14 +4,14 @@ import click
 
 from globus_cli.constants import ExplicitNullType
 from globus_cli.login_manager import LoginManager
-from globus_cli.parsing import command, endpointish_setattr_params
+from globus_cli.parsing import command, endpointish_params
 from globus_cli.termio import Field, TextMode, display
 
 from ._common import deprecated_verify_option
 
 
 @command("mapped", short_help="Create a new GCP Mapped Collection")
-@endpointish_setattr_params("create", name="collection", keyword_style="string")
+@endpointish_params.create(name="collection", keyword_style="string")
 @click.option(
     "--subscription-id",
     help="Set the collection as managed with the given subscription ID",
